edited some buttons for a more clear and clean UI
edited some tabnames

added a dateTimeParseException for date validation

added delete button in the details page for the admin!

added succes/error messages for the admin if an item is deleted or if an error occured

deleted the edit page (it could be implemented BUT is not really necessary and in the requirements)

added validation for the delete product function

added styling/ rules for the students

+++ src/main/java/be/rungroup/eelucaswillaert/service/impl/ProductServiceImpl.java
@@ -51,6 +51,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProduct(long id) {
+        if (!productrepository.existsById(id)) {
+            throw new RuntimeException("Product not found with id: " + id);
+        }
         productrepository.deleteById(id);
 
     }
